Override the directory under which mrgsim.ds creates its temporary directory

By default, mrgsim.ds stores its files under a dedicated temporary directory within tempdir(). Use set_tempdir_base() to specify an alternative location to tempdir().

This is useful in scenarios where the work will be spread across machines (e.g., via Slurm) that do not share access to the file system on which tempdir() resides.

Usage

set_tempdir_base(path)

Arguments

path

The name of an existing directory.

Value

The absolute path (invisibly) to the subdirectory under path that mrgsim.ds will use as its temporary directory.

Details

Although the specified directory may reside anywhere, it should be treated as a temporary directory. Use save_ds() or move_ds() to move the files to more permanent locations.

mrgsim.ds may delete the backing parquet files in its temporary directory (see gc_ds()), but it does not delete path or the temporary directory that it creates within it. The caller is expected to manage this temporary directory, including its removal.

This function aborts if it is called after any mrgsim.ds functionality that relies on the temporary directory.

See also